move dto to req/res and update namespace
