+++ QuizVerse.Application/Core/Interface/IUserService.cs
@@ -1,4 +1,5 @@
-using QuizVerse.Infrastructure.DTOs;
+using QuizVerse.Infrastructure.DTOs.RequestDTOs;
+using QuizVerse.Infrastructure.DTOs.ResponseDTOs;
 
 namespace QuizVerse.Application.Core.Interface;
