+++ QuizVerse.Infrastructure/DTOs/ResponseDTOs/UserDto.cs
@@ -1,4 +1,4 @@
-namespace QuizVerse.Infrastructure.DTOs;
+namespace QuizVerse.Infrastructure.DTOs.ResponseDTOs;
 
 public class UserDto
 {
