+++ QuizVerse.UnitTests/Services/EmailServiceTest.cs
@@ -1,103 +1,7 @@
-// using System.Net.Mail;
-// using System.Threading.Tasks;
-// using Microsoft.Extensions.Configuration;
-// using Moq;
-// using QuizVerse.Application.Core.Service;
-// using QuizVerse.Infrastructure.DTOs;
-// using Xunit;
-
-// namespace QuizVerse.Tests.Services
-// {
-//     public class EmailServiceTests
-//     {
-//         private readonly Mock<IConfiguration> _mockConfig;
-
-//         public EmailServiceTests()
-//         {
-//             _mockConfig = new Mock<IConfiguration>();
-//             _mockConfig.Setup(c => c["EmailSettings:FromAddress"]).Returns("test@domain.com");
-//             _mockConfig.Setup(c => c["EmailSettings:SmtpHost"]).Returns("smtp.domain.com");
-//             _mockConfig.Setup(c => c["EmailSettings:SmtpPort"]).Returns("587");
-//             _mockConfig.Setup(c => c["EmailSettings:SmtpUsername"]).Returns("username");
-//             _mockConfig.Setup(c => c["EmailSettings:SmtpPassword"]).Returns("password");
-//             _mockConfig.Setup(c => c["EmailSettings:EnableSsl"]).Returns("true");
-//         }
-
-//         [Fact]
-//         public async Task SendEmailAsync_ValidRequest_ReturnsTrue()
-//         {
-//             // Arrange
-//             var emailRequest = new EmailRequestDto
-//             {
-//                 To = "recipient@domain.com",
-//                 Subject = "Test Subject",
-//                 Body = "<h1>Hello</h1>"
-//             };
-
-//             var service = new TestableEmailService(_mockConfig.Object);
-
-//             // Act
-//             var result = await service.SendEmailAsync(emailRequest);
-
-//             // Assert
-//             Assert.True(result);
-//             Assert.True(service.SendAsyncCalled); // Custom flag to ensure SendAsync was invoked
-//         }
-
-//         [Fact]
-//         public async Task SendEmailAsync_Exception_ReturnsFalse()
-//         {
-//             // Arrange
-//             var emailRequest = new EmailRequestDto
-//             {
-//                 To = "recipient@domain.com",
-//                 Subject = "Test Subject",
-//                 Body = "<h1>Hello</h1>"
-//             };
-
-//             var service = new TestableEmailService(_mockConfig.Object, throwOnSend: true);
-
-//             // Act
-//             var result = await service.SendEmailAsync(emailRequest);
-
-//             // Assert
-//             Assert.False(result);
-//         }
-
-//         // Custom testable subclass
-//         private class TestableEmailService : EmailService
-//         {
-//             private readonly bool _throwOnSend;
-//             public bool SendAsyncCalled { get; private set; }
-
-//             public TestableEmailService(IConfiguration config, bool throwOnSend = false)
-//                 : base(config)
-//             {
-//                 _throwOnSend = throwOnSend;
-//             }
-
-//             protected override Task SendAsync(SmtpClient client, MailMessage message)
-//             {
-//                 SendAsyncCalled = true;
-
-//                 if (_throwOnSend)
-//                     throw new SmtpException("Simulated failure");
-
-//                 // Do not actually send anything
-//                 return Task.CompletedTask;
-//             }
-//         }
-//     }
-// }
-
-
-
 using System.Net.Mail;
 using Microsoft.Extensions.Configuration;
-using Moq;
 using QuizVerse.Application.Core.Service;
-using QuizVerse.Infrastructure.Common;
-using QuizVerse.Infrastructure.DTOs;
+using QuizVerse.Infrastructure.DTOs.RequestDTOs;
 using Xunit;
 
 namespace QuizVerse.UnitTests.Services;
